+++ wetterdienst/provider/ea/hydrology/api.py
@@ -5,6 +5,7 @@
 from __future__ import annotations
 
 import logging
+from dataclasses import dataclass
 from typing import ClassVar
 
 import polars as pl
@@ -14,7 +15,7 @@ from wetterdienst.core.timeseries.metadata import (
     ParameterModel,
     build_metadata_model,
 )
-from wetterdienst.core.timeseries.request import _DATETIME_TYPE, _PARAMETER_TYPE, _SETTINGS_TYPE, TimeseriesRequest
+from wetterdienst.core.timeseries.request import TimeseriesRequest
 from wetterdienst.core.timeseries.values import TimeseriesValues
 from wetterdienst.metadata.cache import CacheExpiry
 from wetterdienst.util.network import download_file
@@ -209,6 +210,7 @@ class EAHydrologyValues(TimeseriesValues):
         )
 
 
+@dataclass
 class EAHydrologyRequest(TimeseriesRequest):
     """Request class for Environment Agency hydrology data."""
 
@@ -228,29 +230,6 @@ class EAHydrologyRequest(TimeseriesRequest):
         "groundwater_level_max": "level",
         "groundwater_level_min": "level",
     }
-
-    def __init__(
-        self,
-        parameters: _PARAMETER_TYPE,
-        start_date: _DATETIME_TYPE = None,
-        end_date: _DATETIME_TYPE = None,
-        settings: _SETTINGS_TYPE = None,
-    ) -> None:
-        """Initialize the EAHydrologyRequest class.
-
-        Args:
-            parameters: requested parameters
-            start_date: start date of the requested data
-            end_date: end date of the requested data
-            settings: settings for the request
-
-        """
-        super().__init__(
-            parameters=parameters,
-            start_date=start_date,
-            end_date=end_date,
-            settings=settings,
-        )
 
     def _all(self) -> pl.LazyFrame:
         """Acquire all stations and filter for stations that have wanted resolution and parameter combinations."""