+++ wetterdienst/provider/imgw/meteorology/api.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 
 import datetime as dt
 import re
+from dataclasses import dataclass
 from typing import TYPE_CHECKING, ClassVar
 from zoneinfo import ZoneInfo
 
@@ -15,7 +16,7 @@ from dateutil.relativedelta import relativedelta
 from fsspec.implementations.zip import ZipFileSystem
 
 from wetterdienst.core.timeseries.metadata import DatasetModel, build_metadata_model
-from wetterdienst.core.timeseries.request import _DATETIME_TYPE, _PARAMETER_TYPE, _SETTINGS_TYPE, TimeseriesRequest
+from wetterdienst.core.timeseries.request import TimeseriesRequest
 from wetterdienst.core.timeseries.values import TimeseriesValues
 from wetterdienst.metadata.cache import CacheExpiry
 from wetterdienst.metadata.resolution import Resolution
@@ -712,35 +713,13 @@ class ImgwMeteorologyValues(TimeseriesValues):
         return df_files.get_column("url").to_list()
 
 
+@dataclass
 class ImgwMeteorologyRequest(TimeseriesRequest):
     """Request for meteorological data from the Institute of Meteorology and Water Management."""
 
     metadata = ImgwMeteorologyMetadata
     _values = ImgwMeteorologyValues
     _endpoint = "https://dane.imgw.pl/datastore/getfiledown/Arch/Telemetria/Meteo/kody_stacji.csv"
-
-    def __init__(
-        self,
-        parameters: _PARAMETER_TYPE,
-        start_date: _DATETIME_TYPE = None,
-        end_date: _DATETIME_TYPE = None,
-        settings: _SETTINGS_TYPE = None,
-    ) -> None:
-        """Initialize the request.
-
-        Args:
-            parameters: requested parameters
-            start_date: start date
-            end_date: end date
-            settings: settings
-
-        """
-        super().__init__(
-            parameters=parameters,
-            start_date=start_date,
-            end_date=end_date,
-            settings=settings,
-        )
 
     def _all(self) -> pl.LazyFrame:
         """Get all available stations."""